abstract out class for engage generators
